refactor(be): 제품의 가격, 수량을 확인하는데 적절한 데이터 타입으로 교체

+++ backend/src/main/java/com/deliveranything/domain/order/event/dto/OrderItemInfo.java
@@ -5,8 +5,8 @@ import com.deliveranything.domain.order.entity.OrderItem;
 public record OrderItemInfo(
     Long productId,
     String productName,
-    Long quantity,
-    Long price
+    Integer quantity,
+    Integer price
 ) {
 
   public static OrderItemInfo fromOrderItem(OrderItem orderItem) {
